requirements.txt
python-dotenv[cli]==1.0.0
langchain==0.0.137
pinecone-client==2.2.1
ipykernel==6.25.0
openai==0.27.8
tiktoken==0.4.0
ipywidgets==8.0.7
llama-cpp-python==0.1.77
chromadb==0.4.5
